spike-analyzer/requirements.txt
# Core dependencies for Moodle integration
playwright==1.40.0
asyncio==3.4.3

# Configuration and utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Logging and monitoring
structlog==23.2.0

# Data handling
pandas==2.1.4
numpy==1.24.3

# HTTP requests (backup to Playwright)
requests==2.31.0
aiohttp==3.9.1

# JSON and data serialization
ujson==5.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: Database integration (for future stages)
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23

# Optional: AI/ML for content analysis (for future stages)
# openai==1.3.7
# transformers==4.35.2 beautifulsoup4
selenium
